Add some tests and proof of concepts

+++ src/metkit/mars2grib/backend/concepts/concept_core.h
@@ -41,8 +41,8 @@ struct ForEachValue<ValueList<>, Func>
 // ======================================================
 // Constants
 // ======================================================
-static constexpr int NUM_STAGES   = 3;
-static constexpr int NUM_SECTIONS = 5;
+static constexpr uint8_t NUM_STAGES   = 3;
+static constexpr uint8_t NUM_SECTIONS = 5;
 
 // ======================================================
 // Fn — default signature for all table functions
